+++ routes/page.js
@@ -17,6 +17,8 @@ const { postModel }         = require('../models');
 // middleware
 const { _pre, _end }        = require('../configurations/middleware');
 
+// controller
+const { search }            = require('../controllers/search');
 
 
 // ==============================
@@ -27,7 +29,7 @@ router.get('/', (req, res) => res.render('./theme/index'));
 
 
 // searching
-router.get('/search/:TERMS', _pre.doNotCrawled, require('../controllers/search'));
+router.get('/search/:TERMS', _pre.doNotCrawled, search.find);
 
 // router.get('/tag/:TAG', _end.wrapAsync(async (req, res) => {
 //     const result = await postModel.find({ tag: req.params.TAG }).sort({ _created : -1 });
@@ -35,7 +37,7 @@ router.get('/search/:TERMS', _pre.doNotCrawled, require('../controllers/search')
 // }));
 
 // router.get('/:CLASS', _end.wrapAsync(async (req, res) => {
-//     const result = await postModel.find({ class: req.params.CLASS }).sort({ _created : -1 });
+//     const result = await postModel.find({ category: req.params.CLASS }).sort({ _created : -1 });
 //     _end.next.postRender('./theme/search', result)(req, res);
 // }));